Rename `.client` -> `.api`
